+++ qgreenland/exceptions.py
@@ -1,2 +1,6 @@
 class QgrVersionError(Exception):
     pass
+
+
+class QgrInvalidConfigError(Exception):
+    pass