+++ qgreenland/util/vector.py
@@ -2,9 +2,6 @@ import logging
 import os
 import subprocess
 
-import geopandas
-
-
 logger = logging.getLogger('luigi-interface')
 
 
@@ -14,13 +11,6 @@ def cleanup_valid_shapefile(path):
             os.remove(os.path.join(path, f'valid.{ext}'))
         except FileNotFoundError:
             pass
-
-
-def filter_vector(vector_path, *, filter_func):
-    gdf = geopandas.read_file(vector_path)
-
-    gdf = gdf.loc[filter_func]
-    return gdf
 
 
 def ogr2ogr(in_filepath, out_filepath, **ogr2ogr_kwargs):
@@ -33,6 +23,11 @@ def ogr2ogr(in_filepath, out_filepath, **ogr2ogr_kwargs):
         cmd_prefix = f'OGR_ENABLE_PARTIAL_REPROJECTION={enable_partial_reprojection}'
 
     for k, v in ogr2ogr_kwargs.items():
+        # NOTE: Sometimes `v` must be multiple args, e.g. xmin ymin xmax ymax;
+        # Other times, `v` is a string that must be quoted as one argument. We
+        # can't assume which, so the user must do proper quotation in the config
+        # file. This sometimes results in horrible YAML quotation messes like:
+        #     where: '"\"ZONE\" != 0"'
         cmd_args_list.append(f'-{k} {v}')
 
     cmd_args_str = ' '.join(cmd_args_list)